refactor(exports): move export generation into domain use cases
Keep the worker focused on orchestration by moving dataset, trace, and issue export generation into their domain packages. Remove the temporary knip ignores by deleting dead exports and relocating the export tests alongside the new domain use cases.
