+++ apps/workers/src/workers/exports.ts
@@ -1,19 +1,12 @@
 import { promisify } from "node:util"
 import { gzip } from "node:zlib"
-import {
-  csvExportHeader,
-  DatasetRepository,
-  type DatasetRow,
-  DatasetRowRepository,
-  rowsToCsvFragment,
-} from "@domain/datasets"
+import { buildDatasetExportUseCase } from "@domain/datasets"
 import { type EmailSender, exportReadyTemplate, type RenderedEmail, sendEmail } from "@domain/email"
-import { buildExportFilename, type ExportPayload, type ExportSelection } from "@domain/exports"
-import { embedIssueSearchQueryUseCase, IssueProjectionRepository, listIssuesUseCase } from "@domain/issues"
+import type { ExportPayload } from "@domain/exports"
+import { buildIssuesExportUseCase, embedIssueSearchQueryUseCase, IssueProjectionRepository } from "@domain/issues"
 import type { QueueConsumer } from "@domain/queue"
 import {
   DatasetId,
-  DatasetRowId,
   type FilterSet,
   OrganizationId,
   type OrganizationId as OrganizationIdType,
@@ -21,9 +14,8 @@ import {
   type ProjectId as ProjectIdType,
   putInDisk,
   type StorageDiskPort,
-  TraceId,
 } from "@domain/shared"
-import { type Trace, TraceRepository } from "@domain/spans"
+import { buildTracesExportUseCase } from "@domain/spans"
 import { withAi } from "@platform/ai"
 import { AIEmbedLive } from "@platform/ai-voyage"
 import type { ClickHouseClient } from "@platform/db-clickhouse"
@@ -44,7 +36,7 @@ import { IssueProjectionRepositoryLive, withWeaviate } from "@platform/db-weavia
 import { createEmailTransportSender } from "@platform/email-transport"
 import { createStorageDisk } from "@platform/storage-object"
 import { createLogger, withTracing } from "@repo/observability"
-import { Data, Effect, Layer, Schema } from "effect"
+import { Data, Effect, Layer } from "effect"
 import { getClickhouseClient, getPostgresClient, getRedisClient, getWeaviateClient } from "../clients.ts"
 
 const gzipAsync = promisify(gzip)
@@ -55,11 +47,7 @@ class ExportError extends Data.TaggedError("ExportError")<{
 }> {}
 
 const logger = createLogger("exports")
-
-const BATCH_SIZE = 1000
-const ISSUES_EXPORT_BATCH_SIZE = 100
 const SIGNED_URL_EXPIRY_SECONDS = 7 * 24 * 60 * 60
-const traceMetadataFromJsonStringSchema = Schema.fromJsonString(Schema.Record(Schema.String, Schema.String))
 
 const withEmptyIssueProjection = Effect.provide(
   Layer.succeed(IssueProjectionRepository, {
@@ -77,262 +65,15 @@ interface ExportsWorkerDeps {
   emailSender?: EmailSender
 }
 
-/**
- * Compresses CSV content using gzip.
- */
 async function compressCsv(csv: string): Promise<Uint8Array> {
   const compressed = await gzipAsync(csv)
   return new Uint8Array(compressed.buffer, compressed.byteOffset, compressed.byteLength)
 }
 
-/**
- * Generates a dataset export CSV.
- */
-function buildDatasetExportEffect(
-  _organizationId: OrganizationIdType,
-  _projectId: ProjectIdType,
-  datasetId: DatasetId,
-  selection: ExportSelection,
-) {
-  return Effect.gen(function* () {
-    const datasetRepo = yield* DatasetRepository
-    const dataset = yield* datasetRepo.findById(datasetId)
-
-    const rowRepo = yield* DatasetRowRepository
-    const csvChunks: string[] = [csvExportHeader()]
-
-    // First pass: collect all rows if we need to filter by selection
-    if (selection.mode !== "all") {
-      // For selected/allExcept modes, we need to get all rows to filter
-      const allRows: DatasetRow[] = []
-      let offset = 0
-      while (true) {
-        const rows = yield* rowRepo.listPage({
-          datasetId,
-          limit: BATCH_SIZE,
-          offset,
-        })
-        if (rows.length === 0) break
-        allRows.push(...rows)
-        if (rows.length < BATCH_SIZE) break
-        offset += BATCH_SIZE
-      }
-
-      // Apply selection filter
-      const selectedIds =
-        selection.mode === "selected" || selection.mode === "allExcept"
-          ? new Set(selection.rowIds.map((id) => DatasetRowId(id)))
-          : null
-
-      const filteredRows =
-        selection.mode === "selected"
-          ? allRows.filter((r: DatasetRow) => selectedIds?.has(r.rowId))
-          : selection.mode === "allExcept"
-            ? allRows.filter((r: DatasetRow) => !selectedIds?.has(r.rowId))
-            : allRows
-
-      if (filteredRows.length > 0) {
-        csvChunks.push(rowsToCsvFragment(filteredRows))
-      }
-    } else {
-      // For "all" mode, stream in batches
-      let offset = 0
-      while (true) {
-        const rows = yield* rowRepo.listPage({
-          datasetId,
-          limit: BATCH_SIZE,
-          offset,
-        })
-        if (rows.length === 0) break
-        csvChunks.push(rowsToCsvFragment(rows))
-        if (rows.length < BATCH_SIZE) break
-        offset += BATCH_SIZE
-      }
-    }
-
-    const csv = csvChunks.join("")
-    const compressed = yield* Effect.tryPromise({
-      try: () => compressCsv(csv),
-      catch: (e) => new ExportError({ cause: e, kind: "dataset" }),
-    })
-
-    return {
-      content: compressed,
-      filename: buildExportFilename("dataset", dataset.name),
-      exportName: dataset.name,
-    }
-  })
-}
-
-/**
- * Generates a traces export CSV.
- */
-const traceToCsvRow = (trace: Trace): string[] => [
-  trace.traceId,
-  String(trace.spanCount),
-  String(trace.errorCount),
-  trace.startTime.toISOString(),
-  trace.endTime.toISOString(),
-  String(trace.durationNs),
-  String(trace.timeToFirstTokenNs),
-  String(trace.tokensInput),
-  String(trace.tokensOutput),
-  String(trace.tokensCacheRead),
-  String(trace.tokensCacheCreate),
-  String(trace.tokensReasoning),
-  String(trace.tokensTotal),
-  String(trace.costInputMicrocents),
-  String(trace.costOutputMicrocents),
-  String(trace.costTotalMicrocents),
-  trace.sessionId,
-  trace.userId,
-  trace.simulationId || "",
-  trace.tags.join("|"),
-  Schema.encodeSync(traceMetadataFromJsonStringSchema)(trace.metadata),
-  trace.models.join("|"),
-  trace.providers.join("|"),
-  trace.serviceNames.join("|"),
-  trace.rootSpanId,
-  trace.rootSpanName,
-]
-
-export function buildTracesExportEffect(input: {
-  readonly organizationId: OrganizationIdType
-  readonly projectId: ProjectIdType
-  readonly filters?: FilterSet
-  readonly selection?: ExportSelection
-}) {
-  return Effect.gen(function* () {
-    const traceRepo = yield* TraceRepository
-
-    // Use the listing shape columns as the source of truth for v1
-    const headers = [
-      "traceId",
-      "spanCount",
-      "errorCount",
-      "startTime",
-      "endTime",
-      "durationNs",
-      "timeToFirstTokenNs",
-      "tokensInput",
-      "tokensOutput",
-      "tokensCacheRead",
-      "tokensCacheCreate",
-      "tokensReasoning",
-      "tokensTotal",
-      "costInputMicrocents",
-      "costOutputMicrocents",
-      "costTotalMicrocents",
-      "sessionId",
-      "userId",
-      "simulationId",
-      "tags",
-      "metadata",
-      "models",
-      "providers",
-      "serviceNames",
-      "rootSpanId",
-      "rootSpanName",
-    ]
-
-    const csvRows: string[][] = [headers]
-
-    const appendTraces = (traces: readonly Trace[]) => {
-      for (const trace of traces) {
-        csvRows.push(traceToCsvRow(trace))
-      }
-    }
-
-    if (input.selection?.mode === "selected") {
-      const selectedTraceIds = input.selection.rowIds.map((traceId) => TraceId(traceId))
-      const traces = yield* traceRepo.listByTraceIds({
-        organizationId: input.organizationId,
-        projectId: input.projectId,
-        traceIds: selectedTraceIds,
-      })
-      const tracesById = new Map(traces.map((trace) => [trace.traceId, trace] as const))
-
-      for (const traceId of selectedTraceIds) {
-        const trace = tracesById.get(traceId)
-        if (!trace) continue
-
-        if (input.filters) {
-          const matches = yield* traceRepo.matchesFiltersByTraceId({
-            organizationId: input.organizationId,
-            projectId: input.projectId,
-            traceId,
-            filters: input.filters,
-          })
-
-          if (!matches) continue
-        }
-
-        appendTraces([trace])
-      }
-    } else {
-      const excludedTraceIds =
-        input.selection?.mode === "allExcept"
-          ? new Set(input.selection.rowIds.map((traceId) => TraceId(traceId)))
-          : null
-
-      let cursor: { sortValue: string; traceId: string } | undefined
-      while (true) {
-        const page = yield* traceRepo.listByProjectId({
-          organizationId: input.organizationId,
-          projectId: input.projectId,
-          options: {
-            limit: BATCH_SIZE,
-            ...(cursor ? { cursor } : {}),
-            ...(input.filters ? { filters: input.filters } : {}),
-            sortBy: "startTime",
-            sortDirection: "desc",
-          },
-        })
-
-        if (page.items.length === 0) break
-
-        appendTraces(excludedTraceIds ? page.items.filter((trace) => !excludedTraceIds.has(trace.traceId)) : page.items)
-
-        if (!page.hasMore || !page.nextCursor) break
-        cursor = page.nextCursor
-      }
-    }
-
-    // Convert to CSV string
-    const csv = csvRows.map((row) => row.map(escapeCsvField).join(",")).join("\n")
-
-    const compressed = yield* Effect.tryPromise({
-      try: () => compressCsv(csv),
-      catch: (e) => new ExportError({ cause: e, kind: "traces" }),
-    })
-
-    return {
-      content: compressed,
-      filename: buildExportFilename("traces", "project_traces"),
-      exportName: "Project Traces",
-    }
-  })
-}
-
-/**
- * Escapes a field for CSV output.
- */
-function escapeCsvField(value: string): string {
-  // If the value contains commas, quotes, or newlines, wrap it in quotes
-  if (/[",\n\r]/.test(value)) {
-    // Escape quotes by doubling them
-    return `"${value.replace(/"/g, '""')}"`
-  }
-  return value
-}
-
-/**
- * Generates an issues export CSV.
- */
 type IssuesExportInput = {
   readonly organizationId: OrganizationIdType
   readonly projectId: ProjectIdType
-  readonly selection?: ExportSelection
+  readonly selection?: Extract<ExportPayload, { kind: "issues" }>["selection"]
   readonly lifecycleGroup?: "active" | "archived"
   readonly search?: {
     readonly query: string
@@ -366,94 +107,12 @@ const toIssueTimeRange = (
   }
 }
 
-export function buildIssuesExportEffect(input: IssuesExportInput) {
-  return Effect.gen(function* () {
-    const selectionIds =
-      input.selection?.mode === "all" || input.selection === undefined ? null : new Set(input.selection.rowIds)
-    const remainingSelectedIds = input.selection?.mode === "selected" ? new Set(input.selection.rowIds) : null
-
-    // V1 minimal canonical columns per plan
-    const headers = [
-      "id",
-      "uuid",
-      "name",
-      "description",
-      "createdAt",
-      "updatedAt",
-      "escalatedAt",
-      "resolvedAt",
-      "ignoredAt",
-    ]
-
-    const csvRows: string[][] = [headers]
-
-    let offset = 0
-    while (true) {
-      const page = yield* listIssuesUseCase({
-        organizationId: input.organizationId,
-        projectId: input.projectId,
-        limit: ISSUES_EXPORT_BATCH_SIZE,
-        offset,
-        ...(input.lifecycleGroup ? { lifecycleGroup: input.lifecycleGroup } : {}),
-        ...(input.search ? { search: input.search } : {}),
-        ...(input.timeRange ? { timeRange: input.timeRange } : {}),
-        ...(input.sort ? { sort: input.sort } : {}),
-      })
-
-      if (page.items.length === 0) break
-
-      for (const issue of page.items) {
-        if (input.selection?.mode === "selected" && !selectionIds?.has(issue.id)) {
-          continue
-        }
-
-        if (input.selection?.mode === "allExcept" && selectionIds?.has(issue.id)) {
-          continue
-        }
-
-        csvRows.push([
-          issue.id,
-          issue.uuid,
-          escapeCsvField(issue.name),
-          escapeCsvField(issue.description),
-          issue.createdAt.toISOString(),
-          issue.updatedAt.toISOString(),
-          issue.escalatedAt?.toISOString() ?? "",
-          issue.resolvedAt?.toISOString() ?? "",
-          issue.ignoredAt?.toISOString() ?? "",
-        ])
-
-        remainingSelectedIds?.delete(issue.id)
-      }
-
-      if (remainingSelectedIds && remainingSelectedIds.size === 0) break
-      if (!page.hasMore) break
-      offset += page.limit
-    }
-
-    // Convert to CSV string
-    const csv = csvRows.map((row) => row.join(",")).join("\n")
-
-    const compressed = yield* Effect.tryPromise({
-      try: () => compressCsv(csv),
-      catch: (e) => new ExportError({ cause: e, kind: "issues" }),
-    })
-
-    return {
-      content: compressed,
-      filename: buildExportFilename("issues", "project_issues"),
-      exportName: "Project Issues",
-    }
-  })
-}
-
 function generateDatasetExport(
   organizationId: OrganizationIdType,
-  projectId: ProjectIdType,
   datasetId: DatasetId,
-  selection: ExportSelection,
+  selection: Extract<ExportPayload, { kind: "dataset" }>["selection"],
 ) {
-  return buildDatasetExportEffect(organizationId, projectId, datasetId, selection).pipe(
+  return buildDatasetExportUseCase({ datasetId, selection }).pipe(
     withPostgres(DatasetRepositoryLive, getPostgresClient(), organizationId),
     withClickHouse(DatasetRowRepositoryLive, getClickhouseClient(), organizationId),
   )
@@ -463,9 +122,9 @@ function generateTracesExport(
   organizationId: OrganizationIdType,
   projectId: ProjectIdType,
   filters?: FilterSet,
-  selection?: ExportSelection,
+  selection?: Extract<ExportPayload, { kind: "traces" }>["selection"],
 ) {
-  return buildTracesExportEffect({
+  return buildTracesExportUseCase({
     organizationId,
     projectId,
     ...(filters ? { filters } : {}),
@@ -477,7 +136,7 @@ function generateIssuesExport(
   organizationId: OrganizationIdType,
   projectId: ProjectIdType,
   input: {
-    readonly selection?: ExportSelection
+    readonly selection?: Extract<ExportPayload, { kind: "issues" }>["selection"]
     readonly lifecycleGroup?: "active" | "archived"
     readonly searchQuery?: string
     readonly timeRange?: {
@@ -502,7 +161,7 @@ function generateIssuesExport(
 
   const trimmedSearchQuery = input.searchQuery?.trim() || undefined
   if (!trimmedSearchQuery) {
-    return buildIssuesExportEffect(baseEffectInput).pipe(
+    return buildIssuesExportUseCase(baseEffectInput).pipe(
       withPostgres(Layer.mergeAll(EvaluationRepositoryLive, IssueRepositoryLive), getPostgresClient(), organizationId),
       withClickHouse(ScoreAnalyticsRepositoryLive, getClickhouseClient(), organizationId),
       withEmptyIssueProjection,
@@ -520,7 +179,7 @@ function generateIssuesExport(
       catch: (cause) => new ExportError({ cause, kind: "issues" }),
     })
 
-    return yield* buildIssuesExportEffect({
+    return yield* buildIssuesExportUseCase({
       ...baseEffectInput,
       search: {
         query: search.query,
@@ -534,36 +193,25 @@ function generateIssuesExport(
   })
 }
 
-/**
- * Dispatches export generation by kind.
- */
 function dispatchExport(payload: ExportPayload) {
   const organizationId = OrganizationId(payload.organizationId)
   const projectId = ProjectId(payload.projectId)
 
   switch (payload.kind) {
-    case "dataset": {
-      const datasetId = DatasetId(payload.datasetId)
-      return generateDatasetExport(organizationId, projectId, datasetId, payload.selection)
-    }
-    case "traces": {
-      // Cast filters to FilterSet - the queue payload is validated at the boundary
-      const filters = payload.filters as FilterSet | undefined
-      const selection = payload.selection as ExportSelection | undefined
-      return generateTracesExport(organizationId, projectId, filters, selection)
-    }
-    case "issues": {
-      const selection = payload.selection as ExportSelection | undefined
+    case "dataset":
+      return generateDatasetExport(organizationId, DatasetId(payload.datasetId), payload.selection)
+    case "traces":
+      return generateTracesExport(organizationId, projectId, payload.filters, payload.selection)
+    case "issues":
       return generateIssuesExport(organizationId, projectId, {
-        ...(selection ? { selection } : {}),
+        ...(payload.selection ? { selection: payload.selection } : {}),
         ...(payload.lifecycleGroup ? { lifecycleGroup: payload.lifecycleGroup } : {}),
         ...(payload.searchQuery ? { searchQuery: payload.searchQuery } : {}),
         ...(payload.timeRange ? { timeRange: payload.timeRange } : {}),
         ...(payload.sort ? { sort: payload.sort } : {}),
       })
-    }
     default:
-      return Effect.fail(new ExportError({ cause: new Error(`Unknown export kind`) }))
+      return Effect.fail(new ExportError({ cause: new Error("Unknown export kind") }))
   }
 }
 
@@ -584,11 +232,14 @@ export const createExportsWorker = ({
       const projectId = ProjectId(payload.projectId)
 
       return Effect.gen(function* () {
-        // Dispatch to appropriate generator based on kind
         const kind = payload.kind
-        const { content, filename, exportName } = yield* dispatchExport(payload as ExportPayload)
+        const { csv, filename, exportName } = yield* dispatchExport(payload as ExportPayload)
 
-        // Store the compressed artifact
+        const content = yield* Effect.tryPromise({
+          try: () => compressCsv(csv),
+          catch: (cause) => new ExportError({ cause, kind }),
+        })
+
         const fileKey = yield* putInDisk(disk, {
           namespace: "exports",
           organizationId,
@@ -597,7 +248,6 @@ export const createExportsWorker = ({
           filename,
         })
 
-        // Generate signed URL
         const downloadUrl = yield* Effect.tryPromise({
           try: async (): Promise<string> =>
             String(
@@ -605,17 +255,16 @@ export const createExportsWorker = ({
                 expiresIn: SIGNED_URL_EXPIRY_SECONDS,
               }),
             ),
-          catch: (e: unknown) => new ExportError({ cause: e, kind }),
+          catch: (cause: unknown) => new ExportError({ cause, kind }),
         })
 
-        // Send export-ready email
         const rendered = yield* Effect.tryPromise({
           try: (): Promise<RenderedEmail> =>
             exportReadyTemplate({
               exportName,
               downloadUrl,
             }),
-          catch: (e: unknown) => new ExportError({ cause: e, kind }),
+          catch: (cause: unknown) => new ExportError({ cause, kind }),
         })
 
         yield* sendEmailUseCase({
