+++ packages/domain/exports/src/index.ts
@@ -6,9 +6,7 @@ export {
   EXPORT_KINDS,
   type ExportKind,
   type ExportPayload,
-  type ExportProgress,
   type ExportSelection,
-  type ExportWorkflowState,
   exportKindSchema,
   exportPayloadSchema,
   exportSelectionSchema,
